rag_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.39.0
sentence-transformers==3.3.0
faiss-cpu==1.9.0
langdetect==1.0.9
pydantic==2.9.0
python-multipart==0.0.12
aiofiles==24.1.0
numpy>=1.24.0
torch>=2.0.0
